+++ CHANGELOG.md
@@ -5,6 +5,19 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Added
+
+- option `:on_missing_description` on `link` macro which allows to specify 
+  with value `:use_rdf_node` that a linked resource without a description should
+  be kept as an RDF resource (`RDF.IRI` or `RDF.BlankNode`) when preloaded
+  (instead of the default behaviour `:empty_schema` which creating an empty 
+  schema for the linked resource)
+
+[Compare v0.4.1...HEAD](https://github.com/rdf-elixir/grax/compare/v0.4.1...HEAD)
+
+
 ## v0.4.1 - 2023-07-03
 
 ### Fixed